Added generateUuid-function to core
