+++ packages/core/src/index.ts
@@ -27,4 +27,4 @@ export { DvelopContext } from "./context/context";
 export * from "./errors/errors";
 export * from "./http/http-headers";
 export { DvelopHttpRequestConfig, HttpResponse as DvelopHttpResponse, DvelopHttpError, DvelopHttpClient, defaultDvelopHttpClientFactory } from "./http/http-client";
-export { generateRequestId } from "./generate-request-id/generate-request-id";
+export { generateUuid, generateRequestId } from "./generate-uuid/generate-uudi-id";
